\name{glmnetlib-package}
\alias{glmnetlib-package}
\alias{glmnetlib}
\docType{package}
\title{
\packageTitle{glmnetlib}
}
\description{
\packageDescription{glmnetlib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glmnetlib}
\packageIndices{glmnetlib}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{glmnetlib}

Maintainer: \packageMaintainer{glmnetlib}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
