% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirls.r
\name{pirls}
\alias{pirls}
\title{Fit a Penalized Iteratively Reweighted Least Squares (PIRLS) model}
\usage{
pirls(X, y, lambda, alpha = 1, family = binomial, maxit = 500,
  tol = 1e-08, beta = spMatrix(nrow = ncol(X), ncol = 1),
  beta_update = coordinate_descent)
}
\arguments{
\item{X}{the model matrix.}

\item{y}{the response.}

\item{lambda}{the penalty}

\item{alpha}{the elasticnet mixing parameter 0 <= alpha <= 1. Default is
1 (LASSO regression).}

\item{family}{a description of the error distribution and link function
to be used in the model.}

\item{maxit}{the maximum number of iterations for both the weight updates
as well as the beta update. Default is 25.}

\item{tol}{the numeric tolerance. Default is 1e-8.}

\item{beta}{initial slope coefficients. Default is zeros.}

\item{beta_update}{a function for optimizing the slope coefficients
for the current weight matrix. Default is coordinate_descent.}
}
\description{
'pirls' is used to fit the generalized linear model using penalized
maximum likelihood based on the model matrix.
}

