% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_threshold.r
\name{soft_thresh_r}
\alias{soft_thresh_r}
\title{The R implementation of the soft thresholding function.}
\usage{
soft_thresh_r(x, g)
}
\arguments{
\item{x}{a vector of values to soft threshold.}

\item{g}{the threshold paramter}
}
\description{
Soft threshold x with the parameter g in R
}
\examples{
soft_thresh_r(rnorm(10), 0.5)
}

